223: Drop only if table_exists for old, broken migrations

+++ db/migrate/20151218195511_drop_genres.rb
@@ -1,5 +1,5 @@
 class DropGenres < ActiveRecord::Migration
   def change
-    drop_table :genre_movies
+    drop_table :genre_movies if table_exists?(:genre_movies)
   end
 end
